sol-general-api/requirements.txt
bcrypt==4.0.1
bleach==6
Django==4.2.4
django-axes==6.1
django-cleanup==8.0.0
django-cors-headers==4.2.0
django-health-check==3.17
django-http-referrer-policy==1.1.1
django-split-settings==1.2
django-storages==1.13.2
django-stubs-ext==4.2.2
djangorestframework==3.14
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.4
gunicorn==21.2.0
Pillow==10.0
psycopg2-binary==2.9.7
python-decouple==3.8
requests==2.31.0
sqlparse==0.4.4
